+++ lib/gen/dartling_test.dart
@@ -51,7 +51,7 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   sc = '${sc} \n';
 
   sc = '${sc}import "package:unittest/unittest.dart"; \n';
-  sc = '${sc}//import "package:dartling/dartling.dart"; \n';
+  sc = '${sc}import "package:dartling/dartling.dart"; \n';
   sc = '${sc}import "package:${domain.codeLowerUnderscore}_'
        '${model.codeLowerUnderscore}/${domain.codeLowerUnderscore}_'
        '${model.codeLowerUnderscore}.dart"; \n';
@@ -93,6 +93,28 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   sc = '${sc}      model.clear(); \n';
   sc = '${sc}      expect(model.isEmpty, isTrue); \n';
   sc = '${sc}      expect(${entities}.isEmpty, isTrue); \n';
+  sc = '${sc}      expect(${entities}.errors.isEmpty, isTrue); \n';
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("From model to JSON", () { \n';
+  sc = '${sc}      var json = model.toJson(); \n';
+  sc = '${sc}      expect(json, isNotNull); \n';
+  sc = '${sc} \n';
+  sc = '${sc}      print(json); \n';
+  sc = '${sc}      //model.displayJson(); \n';
+  sc = '${sc}      //model.display(); \n';
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("From JSON to model", () { \n';
+  sc = '${sc}      var json = model.toJson(); \n';
+  sc = '${sc}      model.clear(); \n';
+  sc = '${sc}      expect(model.isEmpty, isTrue); \n';
+  sc = '${sc}      model.fromJson(json); \n';
+  sc = '${sc}      expect(model.isEmpty, isFalse); \n';
+  sc = '${sc} \n';
+  sc = '${sc}      model.display(); \n';
   sc = '${sc}    }); \n';
   sc = '${sc} \n';
   
@@ -119,10 +141,8 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   sc = '${sc} \n';
   
   sc = '${sc}    test("Add ${entity} required error", () { \n';
-  var requiredAttribute = findRequiredAttribute(entryConcept);
-  if (requiredAttribute != null) {
-    //var requiredAttributeSet = setAttributeRandomly(requiredAttribute, entity);
-    //sc = '${sc}    ${requiredAttributeSet}';
+  var requiredNonIdAttribute = findRequiredNonIdAttribute(entryConcept);
+  if (requiredNonIdAttribute != null) {
     sc = '${sc}      var ${entity}Concept = ${entities}.concept; \n';
     sc = '${sc}      var ${entity}Count = ${entities}.length; \n';
     sc = '${sc}      var ${entity} = new ${Entity}(${entity}Concept); \n';
@@ -136,7 +156,7 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
     sc = '${sc}      ${entities}.errors.display(title: "Add ${entity} required '
          'error"); \n';
   } else {
-    sc = '${sc}      // no required attribute that is not an id \n';
+    sc = '${sc}      // no required attribute that is not id \n';
   }
   sc = '${sc}    }); \n';
   sc = '${sc} \n';
@@ -145,8 +165,6 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   var idAttribute = findIdAttribute(entryConcept);
   if (idAttribute != null) {
     if (idAttribute.increment == null) {
-      //var idAttributeSet = setAttributeRandomly(idAttribute, entity);
-      //sc = '${sc}    ${idAttributeSet}';
       sc = '${sc}      var ${entity}Concept = ${entities}.concept; \n';
       sc = '${sc}      var ${entity}Count = ${entities}.length; \n';
       sc = '${sc}      var ${entity} = new ${Entity}(${entity}Concept); \n';
@@ -169,39 +187,158 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   sc = '${sc}    }); \n';
   sc = '${sc} \n';
   
-  if (requiredAttribute != null) {
-    sc = '${sc}    test("Find ${entity} by ${requiredAttribute.code}", () { \n';
+  sc = '${sc}    test("Not found ${entity} by new oid", () { \n';
+  sc = '${sc}      var dartlingOid = new Oid.ts(1345648254063); \n';
+  sc = '${sc}      var ${entity} = ${entities}.singleWhereOid(dartlingOid); \n';
+  sc = '${sc}      expect(${entity}, isNull); \n';
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("Find ${entity} by oid", () { \n';
+  sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
+  sc = '${sc}      var ${entity} = '
+       '${entities}.singleWhereOid(random${Entity}.oid); \n';
+  sc = '${sc}      expect(${entity}, isNotNull); \n';
+  sc = '${sc}      expect(${entity}, equals(random${Entity})); \n';
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("Find ${entity} by attribute id", () { \n';
+  if (idAttribute != null) {
     sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
-    sc = '${sc}      var ${entity}${requiredAttribute.codeFirstLetterUpper} = '
-         'random${Entity}.${requiredAttribute.code}; \n';
-    sc = '${sc}      ${Entity} ${entity} = ${entities}.firstWhereAttribute('
-         '"${requiredAttribute.code}", '
-         '${entity}${requiredAttribute.codeFirstLetterUpper}); \n';
+    sc = '${sc}      var ${entity} = \n';
+    sc = '${sc}          ${entities}.singleWhereAttributeId('
+         '"${idAttribute.code}", random${Entity}.${idAttribute.code}); \n';
     sc = '${sc}      expect(${entity}, isNotNull); \n';
+    sc = '${sc}      expect(${entity}.${idAttribute.code}, '
+         'equals(random${Entity}.${idAttribute.code})); \n';
+  } else {
+    sc = '${sc}      // no id attribute \n';
+  }
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("Find ${entity} by required attribute", () { \n';
+  if (requiredNonIdAttribute != null) {
+    sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
+    sc = '${sc}      var ${entity} = \n';
+    sc = '${sc}          ${entities}.firstWhereAttribute('
+         '"${requiredNonIdAttribute.code}", '
+         'random${Entity}.${requiredNonIdAttribute.code}); \n';
+    sc = '${sc}      expect(${entity}, isNotNull); \n';
+    sc = '${sc}      expect(${entity}.${requiredNonIdAttribute.code}, '
+         'equals(random${Entity}.${requiredNonIdAttribute.code})); \n';
     sc = '${sc}    }); \n';
   } else {
-    sc = '${sc}      // Find ${entity} by required attribute: \n';
-    sc = '${sc}      // no required attribute that is not an id \n';
+    sc = '${sc}      // no required attribute that is not id \n';
   }
   sc = '${sc} \n';
   
-  if (requiredAttribute != null) {
-    sc = '${sc}    test("Select ${entities} by ${requiredAttribute.code}", () { \n';
+  sc = '${sc}    test("Find ${entity} by attribute", () { \n';
+  var nonRequiredAttribute = findNonRequiredAttribute(entryConcept);
+  if (nonRequiredAttribute != null) {
     sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
-    sc = '${sc}      var ${entity}${requiredAttribute.codeFirstLetterUpper} = '
-         'random${Entity}.${requiredAttribute.code}; \n';
-    sc = '${sc}      var selected${Entities} = ${entities}.selectWhereAttribute('
-         '"${requiredAttribute.code}", '
-         '${entity}${requiredAttribute.codeFirstLetterUpper}); \n';
+    sc = '${sc}      var ${entity} = \n';
+    sc = '${sc}          ${entities}.firstWhereAttribute('
+         '"${nonRequiredAttribute.code}", random${Entity}.'
+         '${nonRequiredAttribute.code}); \n';
+    sc = '${sc}      expect(${entity}, isNotNull); \n';
+    sc = '${sc}      expect(${entity}.${nonRequiredAttribute.code}, '
+         'equals(random${Entity}.${nonRequiredAttribute.code})); \n';
+  } else {
+    sc = '${sc}      // no attribute that is not required \n';
+  }
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("Select ${entities} by attribute", () { \n';
+  if (nonRequiredAttribute != null) {
+    sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
+    sc = '${sc}      var selected${Entities} = \n';
+    sc = '${sc}          ${entities}.selectWhereAttribute('
+         '"${nonRequiredAttribute.code}", random${Entity}.'
+         '${nonRequiredAttribute.code}); \n';
     sc = '${sc}      expect(selected${Entities}.isEmpty, isFalse); \n';
+    sc = '${sc}      selected${Entities}.forEach((se) => \n';
+    sc = '${sc}          expect(se.${nonRequiredAttribute.code}, '
+         'equals(random${Entity}.${nonRequiredAttribute.code}))); \n';
     sc = '${sc} \n';
     sc = '${sc}      //selected${Entities}.display(title: "Select ${entities} by '
-         '${requiredAttribute.code}"); \n'; 
+         '${nonRequiredAttribute.code}"); \n'; 
+  } else {
+    sc = '${sc}      // no attribute that is not required \n';
+  }
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+
+  sc = '${sc}    test("Select ${entities} by required attribute", () { \n';
+  if (requiredNonIdAttribute != null) {
+    sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
+    sc = '${sc}      var selected${Entities} = \n';
+    sc = '${sc}          ${entities}.selectWhereAttribute('
+         '"${requiredNonIdAttribute.code}", random${Entity}.${requiredNonIdAttribute.code}); \n';
+    sc = '${sc}      expect(selected${Entities}.isEmpty, isFalse); \n';
+    sc = '${sc}      selected${Entities}.forEach((se) => \n';
+    sc = '${sc}          expect(se.${requiredNonIdAttribute.code}, '
+         'equals(random${Entity}.${requiredNonIdAttribute.code}))); \n';
+    sc = '${sc} \n';
+    sc = '${sc}      //selected${Entities}.display(title: "Select ${entities} by '
+         '${requiredNonIdAttribute.code}"); \n'; 
     sc = '${sc}    }); \n';
   } else {
-    sc = '${sc}      // Select ${entities} by required attribute: \n';
-    sc = '${sc}      // no required attribute that is not an id \n';
+    sc = '${sc}      // no required attribute that is not id \n';
   }
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("Select ${entities} by attribute, then add", () { \n';
+  var nonIdAttribute = findNonIdAttribute(entryConcept);
+  if (nonIdAttribute != null) {
+    sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
+    sc = '${sc}      var selected${Entities} = \n';
+    sc = '${sc}          ${entities}.selectWhereAttribute('
+         '"${nonIdAttribute.code}", random${Entity}.${nonIdAttribute.code}); \n';
+    sc = '${sc}      expect(selected${Entities}.isEmpty, isFalse); \n';
+    sc = '${sc}      expect(selected${Entities}.source.isEmpty, isFalse); \n';
+    sc = '${sc}      var ${entities}Count = ${entities}.length; \n';
+    sc = '${sc} \n';
+    sc = '${sc}      var ${entity} = new ${Entity}(${entities}.concept); \n';
+    var attributesSet = setAttributesRandomly(entryConcept, entity);
+    sc = '${sc}${attributesSet}';
+    sc = '${sc}      var added = selected${Entities}.add(${entity}); \n';
+    sc = '${sc}      expect(added, isTrue); \n';
+    sc = '${sc}      expect(${entities}.length, equals(++${entities}Count)); \n';
+    sc = '${sc} \n';
+    sc = '${sc}      //selected${Entities}.display(title: \n';
+    sc = '${sc}      //  "Select ${entities} by attribute, then add"); \n';
+    sc = '${sc}      //${entities}.display(title: "All ${entities}"); \n'; 
+  } else {
+    sc = '${sc}      // no attribute that is not id \n';
+  }
+  sc = '${sc}    }); \n';
+  sc = '${sc} \n';
+  
+  sc = '${sc}    test("Select ${entities} by attribute, then remove", () { \n';
+  if (nonIdAttribute != null) {
+    sc = '${sc}      var random${Entity} = ${entities}.random(); \n';
+    sc = '${sc}      var selected${Entities} = \n';
+    sc = '${sc}          ${entities}.selectWhereAttribute('
+         '"${nonIdAttribute.code}", random${Entity}.${nonIdAttribute.code}); \n';
+    sc = '${sc}      expect(selected${Entities}.isEmpty, isFalse); \n';
+    sc = '${sc}      expect(selected${Entities}.source.isEmpty, isFalse); \n';
+    sc = '${sc}      var ${entities}Count = ${entities}.length; \n';
+    sc = '${sc} \n';
+    sc = '${sc}      var removed = selected${Entities}.remove(random${Entity}); \n';
+    sc = '${sc}      expect(removed, isTrue); \n';
+    sc = '${sc}      expect(${entities}.length, equals(--${entities}Count)); \n';
+    sc = '${sc} \n';
+    sc = '${sc}      random${Entity}.display(prefix: "removed"); \n';
+    sc = '${sc}      //selected${Entities}.display(title: \n';
+    sc = '${sc}      //  "Select ${entities} by attribute, then remove"); \n';
+    sc = '${sc}      //${entities}.display(title: "All ${entities}"); \n'; 
+  } else {
+    sc = '${sc}      // no attribute that is not id \n';
+  }
+  sc = '${sc}    }); \n';
   sc = '${sc} \n';
   
   sc = '${sc}    test("Sort ${entities}", () { \n';
@@ -270,10 +407,10 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   sc = '${sc} \n';
   
   sc = '${sc}    test("True for every ${entity}", () { \n';
-  if (requiredAttribute != null) {
-    sc = '${sc}      expect(${entities}.every((e) => e.${requiredAttribute.code} != null), isTrue); \n';
+  if (requiredNonIdAttribute != null) {
+    sc = '${sc}      expect(${entities}.every((e) => e.${requiredNonIdAttribute.code} != null), isTrue); \n';
   } else {
-    sc = '${sc}      // no required attribute that is not an id \n';
+    sc = '${sc}      // no required attribute that is not id \n';
   }
   sc = '${sc}    }); \n';
   sc = '${sc} \n';
@@ -301,7 +438,25 @@ String genDartlingTest(Repo repo, Model model, Concept entryConcept) {
   return sc;
 }
 
-Attribute findRequiredAttribute(Concept concept) {
+Attribute findNonIdAttribute(Concept concept) {
+  for (Attribute attribute in concept.attributes) {
+    if (!attribute.identifier) {
+      return attribute;
+    }
+  }
+  return null;
+}
+
+Attribute findNonRequiredAttribute(Concept concept) {
+  for (Attribute attribute in concept.attributes) {
+    if (!attribute.required) {
+      return attribute;
+    }
+  }
+  return null;
+}
+
+Attribute findRequiredNonIdAttribute(Concept concept) {
   for (Attribute attribute in concept.attributes) {
     if (attribute.required && !attribute.identifier) {
       return attribute;
@@ -319,30 +474,6 @@ Attribute findIdAttribute(Concept concept) {
   return null;
 }
 
-/*
-String setAttributeRandomly(Attribute attribute, String entity) {
-  var sc = '';
-  if (attribute.type.code == 'String') {     
-    sc = '${sc}    ${entity}.${attribute.code} = "${randomWord()}"; \n';      
-  }  else if (attribute.type.code == 'num') {
-    sc = '${sc}  ${entity}.${attribute.code} = ${randomNum(1000)}; \n';
-  } else if (attribute.type.code == 'int') {
-    sc = '${sc}  ${entity}.${attribute.code} = ${randomInt(10000)}; \n';
-  } else if (attribute.type.code == 'double') {
-    sc = '${sc}  ${entity}.${attribute.code} = ${randomDouble(100)}; \n';
-  } else if (attribute.type.code == 'bool') {
-    sc = '${sc}  ${entity}.${attribute.code} = ${randomBool()}; \n';
-  } else if (attribute.type.code == 'DateTime') {
-    sc = '${sc}  ${entity}.${attribute.code} = new DateTime.now(); \n';
-  } else if (attribute.type.code == 'Uri') {
-    sc = '${sc}  ${entity}.${attribute.code} = Uri.parse("${randomUri()}"); \n';
-  } else if (attribute.type.code == 'Email') {
-    sc = '${sc}  ${entity}.${attribute.code} = ${randomEmail()}); \n';
-  } else {
-    sc = '${sc}  ${entity}.${attribute.code} = ${randomWord()}; \n';
-  }
-  return sc;
-}
-*/
+