+++ CHANGELOG.md
@@ -4,6 +4,7 @@ based on [Semantic Versioning](http://semver.org/)
 
 *2.0.0* 2014-03-15
 
++ handle id attribute increment in add propagations
 + order entry concepts for model init 
 + drop increment attribute in model init; 
   drop increment attribute id unique error test